Gestion Stream + affichage des unités
Back to past: since the use of process don't work with windows and
freezing (freeze open an infinity of subprocess which make the
application useless), we choose to come back to thread.
Thread is working well

plus: (mistake forget to make two different commit)
the mu symbol now is correctly print on plot

Bug: when closing the applications, sometimes thread is still running in
the background, but it's still better than stream not working at all.

improve: the way i fix the mu symbol printing can be improve i think